【その他】Menuで使われているatom or moleculesのパッケージングを修正 #230
パッケージングについてリファクタをしています。
