+++ front/src/components/molecules/menu/MenuIconClickableUnit.tsx
@@ -1,8 +1,10 @@
+'use client';
+
 import { LucideIconType } from "@/types/ui/icon";
 import { cn } from "@/lib/shadcn/utils";
-import MenuIcon from "@/components/atoms/MenuIcon";
-import MenuTitle from "@/components/atoms/MenuTitle";
-import Link from "next/link";
+import MenuIcon from "@/components/atoms/menu/MenuIcon";
+import MenuTitle from "@/components/atoms/menu/MenuTitle";
+import { useState } from "react";
 
 /**
  * メニューの行を表示するコンポーネントのprops
@@ -10,11 +12,10 @@ import Link from "next/link";
 interface Props {
   title: string;
   icon?: LucideIconType;
-  href: string;
+  onClick: () => Promise<void>;
   divClassName?: string | string[];
   iconClassName?: string | string[];
   titleClassName?: string | string[];
-  targetBlank?: boolean;
 }
 
 /**
@@ -23,12 +24,11 @@ interface Props {
 const defaultDivClassName = ["flex", "items-center", "gap-2", "hover:bg-gray-100", "rounded-lg", "transition-colors", "cursor-pointer", "p-2"];
 
 /**
- * メニューの行を表示するコンポーネント（リンク付き）
+ * メニューの行を表示するコンポーネント（クリック可能）
  * 
- * このコンポーネントは、アイコンとタイトルを横並びで表示するメニュー行を作成し、クリック可能なリンクとして機能します。
+ * このコンポーネントは、アイコンとタイトルを横並びで表示するメニュー行を作成し、クリック可能なボタンとして機能します。
  * デフォルトでは、アイコンは40x40pxのサイズで表示され、タイトルはアイコンの右側に配置されます。
  * ホバー時には背景色が変化し、視覚的なフィードバックを提供します。
- * リンクは別窓で開きます。
  *  
  * スタイルのカスタマイズ:
  * - divClassName: メニュー行全体のスタイルをカスタマイズします（例：背景色、パディング、ホバー効果など）
@@ -37,10 +37,10 @@ const defaultDivClassName = ["flex", "items-center", "gap-2", "hover:bg-gray-100
  * 
  * 使用例:
  * ```tsx
- * <MenuIconLinkUnit
+ * <MenuIconClickableUnit
  *   icon={User}
  *   title="ユーザー設定"
- *   href="/settings"
+ *   onClick={() => console.log("clicked")}
  *   divClassName={["hover:bg-blue-100"]}
  *   iconClassName={["text-blue-500"]}
  *   titleClassName={["font-bold"]}
@@ -49,28 +49,45 @@ const defaultDivClassName = ["flex", "items-center", "gap-2", "hover:bg-gray-100
  * 
  * @param icon - 表示するアイコン（Lucideアイコン）
  * @param title - 表示するタイトルテキスト
- * @param href - リンク先のURL
+ * @param onClick - クリック時のコールバック関数
  * @param divClassName - メニュー行のコンテナに適用する追加のクラス名
  * @param iconClassName - アイコンに適用する追加のクラス名
  * @param titleClassName - タイトルに適用する追加のクラス名
- * @param targetBlank - リンク先を別窓で開くかどうか（デフォルトはtrue）
- * @returns リンク付きメニュー行のコンポーネント
+ * @returns クリック可能なメニュー行のコンポーネント
  */
-export default function MenuIconLinkUnit({ 
+export default function MenuIconClickableUnit({ 
   icon, 
   title, 
-  href,
+  onClick,
   divClassName,
   iconClassName,
-  titleClassName,
-  targetBlank = true
+  titleClassName 
 }: Readonly<Props>) {
+  const [isPending, setIsPending] = useState(false);
+
+  const handleClick = async () => {
+    if (isPending) return;
+    setIsPending(true);
+    try {
+      await onClick();
+    } finally {
+      setIsPending(false);
+    }
+  };
+
   return (
-    <Link href={href} className="block" target={targetBlank ? "_blank" : "_self"} rel="noopener noreferrer">
-      <div className={cn(defaultDivClassName, divClassName)}>
-        {icon && <MenuIcon icon={icon} className={iconClassName} />}
-        <MenuTitle title={title} className={titleClassName} />
-      </div>
-    </Link>
+    <button 
+      className={cn(
+        defaultDivClassName,
+        isPending && "opacity-50 cursor-not-allowed",
+        divClassName
+      )}
+      onClick={handleClick}
+      disabled={isPending}
+      type="button"
+    >
+      {icon && <MenuIcon icon={icon} className={iconClassName} />}
+      <MenuTitle title={title} className={titleClassName} />
+    </button>
   )
 }
