+++ front/src/components/organisms/sample/atomic/SampleAtomic2.tsx
@@ -1,5 +1,5 @@
 import { User } from "lucide-react";
-import MenuIconUnit from "@/components/molecules/MenuIconUnit";
+import MenuIconUnit from "@/components/molecules/menu/MenuIconUnit";
 
 export default function SampleAtomic2() {
   return (
