+++ front/src/components/organisms/sample/atomic/SampleAtomic4.tsx
@@ -1,5 +1,5 @@
 import { Bell } from "lucide-react";
-import MenuIconUnit from "@/components/molecules/MenuIconUnit";
+import MenuIconUnit from "@/components/molecules/menu/MenuIconUnit";
 
 export default function SampleAtomic4() {
   return (
